Use better axis label step size
The distance between axis ticks and labels is considered nice if it is a multiple of 10 raised to appropriate power, i.e. 1, 2, 5 or 10. Previous implementation only used powers of 10, i.e. 0.1, 1, 10 etc, which made the number of ticks and steps jump a lot. Using 0.1, 0.2, 0.5, 1, 2, 5, 10, 20, etc insted produces a much smoother plot, especially when zooming in and out.

Furthermore, the way the distances between the labels was handled made the labels jump all over just by paning due to how the labels were filtered away since labels was not shown on every tick. Better to add a label to every tick so that the labels doesn't change place as you pan around.

The new implementation also makes sure that we have at least 50 pixels center distance between each tick or 20 pixels separation between each label on the X-axis since they can get pretty wide