Resizing the frame moves the plot
If you zoom a series and then resize the frame, the series will move away back to a previous position. This is due to that OnSize uses m_desired to redraw the plot, but in many instances of zooming the m_desired was not updated correctly since the margins was not considered, or was not updated at all. Fixed by recalculating m_desired whenever m_posX, m_scaleX, m_posY or m_scaleY is updated and also taking margins into account.