Add support for multiple Y-axes
Added support for an arbitrary number of Y-axes. This mostly comes down to remaking m_YAxis and m_Y2Axis into a m_YAxisList vector, as well as m_scaleY, m_scaleY2, m_posY and m_posY2 into another m_yAxisDataList vector. Also the mpFloatRect that keeps the boundaries of the plots need to support multiple Y-axis

Furthermore, if using multiple Y-axis they shouldn't be drawn on top of each other. To avoid this but still keeping the idea of adjustable margins,  an additional variable m_marginOuter is added to handle the margin exluding the Y-axes, and let the existing m_margin include the Y-axis width so that the plotting area is drawn correctly without plotting on top of the various Y-axes. This will still let you adjust the "outer" margin without worrying about the Y-axes width.

To use multiple Y-axis, you assign a Y-axis index your mpFunction in their constructor, and in the same way you can add a Y-axis (mpScaleY) with a specific index which will then be connected to all plots using that index. It is not possible to add multiple mpScaleY with the same index to AddLayer, but you can add several mpFunction's that points to the same Y-axis index.

In configuration, you can now assign a Serie or Line to a Y-axis via a choice dropdown list, and you can add as many Y-axis as you want.