Auto update Y-axis width
Depending on label and name size, the Y-axis width can differ and need to by updated dynamically as you zoom or change font on the axis.To calculate width, the label and label format need to be known, which currently only is calculated in OnPaint->DoPlot, but we don't want to update axis width in a OnPaint event. Instead move the calculation of label format to a function and us it in UpdateAll() to make sure that axis width is up to date before the OnPaint event is triggered,

Now that we automatically adjust the axis width, there's no reason to draw the axis name horizontally, thus always point it vertically. At least for mpALIGN_LEFT and mpALIGN_RIGHT.

Also a minor change regarding the extra margin since the first Y-axis ends up outside of the plot otherwise