Add demo for multiply Y-axis
The demo for multiply Y-axis shows the same plots as for "Draw Sample" but with each function on separate Y-axis. All Y-axis is on the left since I think that looks best, and the colors and width of the axis has also been adjusted slightly to yield a better look for multiple Y-axis, Also a small tool tip is shown when you click the Multi Y-axis button to describe how to interact with individual Y-axis regarding zoom and panning.

Also added a GetLayerYAxis() function to mathplot in order to easier delete the Y-axis when trying a different demo plot