Better calculation of Y-axis width and Y-axis mpInfoCoords
The Y-axis width tries to estimate the maximum size of the labels by taking the uppermost and lowermost value and checking their text size. But it did not account for that the labels are "rounded" to a specific step size, thus a lot of the time the %g format specifier added a lot of decimals which will never be shown in this specific zoom level. Fixed so that the upper and lower value is rounded properly according to the label step size. Also corrected placement of axis name to be consistant with the axis width and added constants for tick size and extra axis spacing to reduce the number of hard-coded values in the code.

Fixed so that the Y-axis mpInfoCoords only shows Y-axis values if the Y-axis exists or any function actually uses a specific Y-axis so that if you remove an Y-axis or a function it shall be reflected in the mpInfoCoords. Needed since the m_yAxisDataList that stores scales and position can never be deleted or reduced since it relys on that the index in its vector correspond to a specific Y-axis. Also added Y-axis name to the mpInfoCoords

Fixed so that Y-axis name is indexed from 0 instead of 1 in MathPlotConfig since it is indexed from 0 in mpInfoCoords and better corresponds to the actual index