Restructure mpFloatRect to group vectors
To avoid the paralell vectors YminList and YmaxList it is better to group them into a struct, so that their size is consistent. Created new mpRange struct for this containing min and max. Also grouped the Xmin and Xmax into mpRange to be consistent with the Y values. With this restructuring a few of the for loop could be simplified to ranged for loop. Some for loop stil need index-based for loops when needing to work with both m_yAxisDataList and mpFloatRect values.

Also fixed some int to size_t type for Y-indices variables