Fix y-size of mpFloatRect
Currently the y-size of mpFloatRect is stored in a static variable so that when you create a new instance the size stays the same as existing instances. This does not work when you have multiple mpWindows as they can have different number of y-axis. Instead the y-size of mpFloatRect must match the number of y-axis of its parent mpWindow. In order to ensure that these sizes match, only allow a constructor which pass a mpWindow as argument so that the y-size can be set automatically via the mpWindow instance.