Correction for min/max calculation in mpFXYVector
For some specific mpFXYVector series the min max calculation becomes incorrect. E.g. given a time series with increasing X-values, if the first two points in a series is much farther apart then the rest of the points, the first calculation of m_deltaX can get too large, resulting in a too large max value. And, for increasing X-values, the m_deltaX is only added to the max value and not subtracted from the min value, yielding unsymmetric boundaries. Furthermore, if bars are used, the leftmost bar (or sometimes both) will be drawn halfway outside the plot.

It seems to me that the delta is only needed for the X-axis if you plot bars, and then it's only necessary to add half of the delta to the min and max value in order to make the outermost bars visible. And only the minimum delta should be added/subtracted to the max/min value, and in order to do so consistently it need to be added after all of the points have been calculated. Fixed this by only adding it in the Getter() functions, so that the min/max variables ir "pure" min/max value and don't have to include the delta if it's not needed